chore(memory: vmm): move page_faul handler inside vmm'code
